requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-cors-headers==4.3.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7